+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/fasterroute/FasterRouteController.kt
@@ -3,12 +3,14 @@ package com.mapbox.navigation.core.fasterroute
 import com.mapbox.api.directions.v5.models.DirectionsRoute
 import com.mapbox.api.directions.v5.models.RouteOptions
 import com.mapbox.navigation.base.extensions.ifNonNull
+import com.mapbox.navigation.core.directions.session.AdjustedRouteOptionsProvider
 import com.mapbox.navigation.core.directions.session.DirectionsSession
 import com.mapbox.navigation.core.directions.session.RoutesRequestCallback
 import com.mapbox.navigation.core.trip.session.TripSession
 import com.mapbox.navigation.utils.timer.MapboxTimer
 
 internal class FasterRouteController(
+    private val adjustedRouteObjectsProvider: AdjustedRouteOptionsProvider,
     private val directionsSession: DirectionsSession,
     private val tripSession: TripSession
 ) {
@@ -31,12 +33,9 @@ internal class FasterRouteController(
     }
 
     private fun requestFasterRoute() {
-        ifNonNull(
-            directionsSession.getRouteOptions(),
-            tripSession.getEnhancedLocation()
-        ) { options, enhancedLocation ->
-            val routeProgress = tripSession.getRouteProgress() ?: return
-            val optionsRebuilt = directionsSession.getAdjustedRouteOptions(options, routeProgress, enhancedLocation)
+        ifNonNull(tripSession.getEnhancedLocation()) { enhancedLocation ->
+            val optionsRebuilt = adjustedRouteObjectsProvider.getRouteOptions(enhancedLocation)
+                ?: return
             directionsSession.requestFasterRoute(optionsRebuilt, fasterRouteRequestCallback)
         }
     }
